bot/requirements.txt
aiogram==3.20.0.post0
python-dotenv ==1.1.0
httpx ==0.28.1


numpy>=1.23.0
tensorflow>=2.12.0


matplotlib>=3.5.0


pandas>=2.0.0

influxdb-client